requirements.txt
colorama
pyyaml
latex